Add Client connection for post requests
Adds the Braze API client for post requests.
Uses faraday to connect to the Braze API and make
a post request when instantiated with an api_key, app_id,
and braze_url.
Parses the response and raises the relevant error if necessary
using the ResponseParser class.
Adds spec coverage for the Client and the ResponseParser.